Update config format, wrap Vector3D in AT_Vector3D for more effective minification and obfuscation

+++ ArgusV2/Helper/Math.cs
@@ -6,7 +6,7 @@ namespace IngameScript.Helper
     {
         public static double SnapToMultiple(double value, double multiple)
         {
-            return (double)(Math.Round(value / multiple) * multiple);
+            return (Math.Round(value / multiple) * multiple);
         }
     }
 }
